+++ src/constants/prop-types/related.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import predefinedPropTypes from './img-with-placeholder'
 
-const card = PropTypes.shape({
+const card = {
   category: PropTypes.string,
   date: PropTypes.string,
   desc: PropTypes.string,
@@ -10,8 +10,9 @@ const card = PropTypes.shape({
   isTargetBlank: PropTypes.bool,
   thumbnail: predefinedPropTypes.imagePropType,
   title: PropTypes.string.isRequired,
-})
+}
 
 export default {
   card,
+  data: PropTypes.arrayOf(PropTypes.shape(card)),
 }